fix(i18next): move module augmentations to `index` and import in `register`
